requirements.txt
# Core
sentence-transformers==3.0.1
chromadb==0.5.5
rank-bm25==0.2.2

# Document loading
pypdf==4.3.1
python-docx==1.1.2

# LLM providers
openai==1.40.0
anthropic==0.34.2

# Utilities
python-dotenv==1.0.1
tqdm==4.66.5
nltk==3.9.1
tiktoken==0.7.0
rich==13.7.1          # Pretty CLI output

# Testing
pytest==8.3.2
pytest-mock==3.14.0
